[Chore](nereids) Remove CopyStmt
