+++ fe/fe-core/src/main/cup/sql_parser.cup
@@ -2004,18 +2004,6 @@ job_label ::=
     :}
     ;
 
-// Copy Statement
-copy_stmt ::=
-    KW_COPY KW_INTO opt_select_hints:hints table_name:name opt_col_list:cols KW_FROM copy_from_param:copyFromParam KW_PROPERTIES opt_key_value_map:properties
-    {:
-        RESULT = new CopyStmt(name, cols, copyFromParam, properties, hints);
-    :}
-    | KW_COPY KW_INTO opt_select_hints:hints table_name:name opt_col_list:cols KW_FROM copy_from_param:copyFromParam
-    {:
-        RESULT = new CopyStmt(name, cols, copyFromParam, new HashMap(), hints);
-    :}
-    ;
-
 copy_from_param ::=
     stage_and_pattern:stage
     {:
