+++ fe/fe-core/src/main/java/org/apache/doris/qe/DdlExecutor.java
@@ -23,7 +23,6 @@ import org.apache.doris.analysis.AlterSqlBlockRuleStmt;
 import org.apache.doris.analysis.AlterTableStmt;
 import org.apache.doris.analysis.CancelExportStmt;
 import org.apache.doris.analysis.CancelLoadStmt;
-import org.apache.doris.analysis.CopyStmt;
 import org.apache.doris.analysis.CreateEncryptKeyStmt;
 import org.apache.doris.analysis.CreateIndexPolicyStmt;
 import org.apache.doris.analysis.CreateJobStmt;
@@ -49,22 +48,12 @@ import org.apache.doris.analysis.SyncStmt;
 import org.apache.doris.analysis.UninstallPluginStmt;
 import org.apache.doris.catalog.EncryptKeyHelper;
 import org.apache.doris.catalog.Env;
-import org.apache.doris.cloud.load.CloudLoadManager;
-import org.apache.doris.cloud.load.CopyJob;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.job.exception.JobException;
-import org.apache.doris.load.EtlStatus;
-import org.apache.doris.load.FailMsg;
-import org.apache.doris.load.loadv2.JobState;
-import org.apache.doris.load.loadv2.LoadJob;
 
-import com.google.common.collect.Lists;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
-import java.util.List;
-import java.util.Map;
 
 /**
  * Use for execute ddl.
@@ -150,94 +139,9 @@ public class DdlExecutor {
             RefreshCatalogStmt refreshCatalogStmt = (RefreshCatalogStmt) ddlStmt;
             env.getRefreshManager()
                     .handleRefreshCatalog(refreshCatalogStmt.getCatalogName(), refreshCatalogStmt.isInvalidCache());
-        } else if (ddlStmt instanceof CopyStmt) {
-            executeCopyStmt(env, (CopyStmt) ddlStmt);
         } else {
             LOG.warn("Unkown statement " + ddlStmt.getClass());
             throw new DdlException("Unknown statement.");
-        }
-    }
-
-    public static void executeCopyStmt(Env env, CopyStmt copyStmt) throws Exception {
-        CopyJob job = (CopyJob) (((CloudLoadManager) env.getLoadManager()).createLoadJobFromStmt(copyStmt));
-        if (!copyStmt.isAsync()) {
-            // wait for execute finished
-            waitJobCompleted(job);
-            if (job.getState() == JobState.UNKNOWN || job.getState() == JobState.CANCELLED) {
-                QueryState queryState = new QueryState();
-                FailMsg failMsg = job.getFailMsg();
-                EtlStatus loadingStatus = job.getLoadingStatus();
-                List<List<String>> result = Lists.newArrayList();
-                List<String> entry = Lists.newArrayList();
-                entry.add(job.getCopyId());
-                entry.add(job.getState().toString());
-                entry.add(failMsg == null ? "" : failMsg.getCancelType().toString());
-                entry.add(failMsg == null ? "" : failMsg.getMsg());
-                entry.add("");
-                entry.add("");
-                entry.add("");
-                entry.add(loadingStatus.getTrackingUrl());
-                result.add(entry);
-                queryState.setResultSet(new ShowResultSet(copyStmt.getMetaData(), result));
-                ConnectContext.get().setState(queryState);
-                return;
-            } else if (job.getState() == JobState.FINISHED) {
-                EtlStatus loadingStatus = job.getLoadingStatus();
-                Map<String, String> counters = loadingStatus.getCounters();
-                QueryState queryState = new QueryState();
-                List<List<String>> result = Lists.newArrayList();
-                List<String> entry = Lists.newArrayList();
-                entry.add(job.getCopyId());
-                entry.add(job.getState().toString());
-                entry.add("");
-                entry.add("");
-                entry.add(counters.getOrDefault(LoadJob.DPP_NORMAL_ALL, "0"));
-                entry.add(counters.getOrDefault(LoadJob.DPP_ABNORMAL_ALL, "0"));
-                entry.add(counters.getOrDefault(LoadJob.UNSELECTED_ROWS, "0"));
-                entry.add(loadingStatus.getTrackingUrl());
-                result.add(entry);
-                queryState.setResultSet(new ShowResultSet(copyStmt.getMetaData(), result));
-                ConnectContext.get().setState(queryState);
-                return;
-            }
-        }
-        QueryState queryState = new QueryState();
-        List<List<String>> result = Lists.newArrayList();
-        List<String> entry = Lists.newArrayList();
-        entry.add(job.getCopyId());
-        entry.add(job.getState().toString());
-        entry.add("");
-        entry.add("");
-        entry.add("");
-        entry.add("");
-        entry.add("");
-        entry.add("");
-        result.add(entry);
-        queryState.setResultSet(new ShowResultSet(copyStmt.getMetaData(), result));
-        ConnectContext.get().setState(queryState);
-    }
-
-    private static void waitJobCompleted(CopyJob job) throws InterruptedException {
-        // check the job is completed or not.
-        // sleep 10ms, 1000 times(10s)
-        // sleep 100ms, 1000 times(100s + 10s = 110s)
-        // sleep 1000ms, 1000 times(1000s + 110s = 1110s)
-        // sleep 5000ms...
-        long retry = 0;
-        long currentInterval = 10;
-        while (!job.isCompleted()) {
-            Thread.sleep(currentInterval);
-            if (retry > 3010) {
-                continue;
-            }
-            retry++;
-            if (retry > 3000) {
-                currentInterval = 5000;
-            } else if (retry > 2000) {
-                currentInterval = 1000;
-            } else if (retry > 1000) {
-                currentInterval = 100;
-            }
         }
     }
